requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2

# PDF processing
PyPDF2==3.0.1

# Machine learning and embeddings
sentence-transformers==2.2.2
huggingface_hub==0.16.4
torch>=1.9.0
transformers>=4.21.0

# Vector search
faiss-cpu==1.8.0
numpy>=1.21.0

# Data models and validation
pydantic==2.5.0

# Optional: Enhanced answer generation
openai==1.3.0

# Security (if needed for future features)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.0  # For testing FastAPI endpoints

# Development tools
black==23.9.1  # Code formatting
isort==5.12.0  # Import sorting